ref(ui): hide team list & team invites behind a dropdown in the nav
We hardly ever use these so might as well hide them away.
