+++ frontend/src/components/scss/helpers.scss
@@ -203,16 +203,11 @@
   grid-row: 1;
 }
 
-.max-width-200px {
-  max-width: 200px;
-}
-
-.max-width-400px {
-  max-width: 400px;
-}
-
-.mw-1000px {
-  max-width: 1000px !important;
+@each $i in (200px, 400px, 800px, 1000px) {
+  .mw-#{$i},
+  .max-width-#{$i} {
+    max-width: $i !important;
+  }
 }
 
 @for $i from 0 through 16 {
